# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.CategoryResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.ProductCreate:
    properties:
      category_id:
        type: integer
      code:
        type: string
      description:
        type: string
      name:
        type: string
      price:
        type: number
    required:
    - category_id
    - code
    - name
    - price
    type: object
  schemas.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: number
    type: object
  schemas.ProductResponseDTO:
    properties:
      category:
        $ref: '#/definitions/schemas.CategoryResponse'
      code:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      stock:
        type: number
    type: object
  schemas.ProductUpdate:
    properties:
      category_id:
        type: integer
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    required:
    - category_id
    - code
    - id
    - name
    - price
    type: object
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  schemas.RoleResponse:
    properties:
      id:
        type: integer
      name:
        enum:
        - admin
        - vendedor
        - repositor
        type: string
    type: object
...
info:
  contact: {}
  description: This is a API for Club Norte
  termsOfService: http://swagger.io/terms/
  title: APP Club Norte
  version: "1.0"
paths:
...
  /v1/product/create:
    post:
      consumes:
      - application/json
      description: ProductCreate crear nuevo producto
      parameters:
      - description: Información del producto a crear
        in: body
        name: productCreate
        required: true
        schema:
          $ref: '#/definitions/schemas.ProductCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductCreate
      tags:
      - Product
  /v1/product/delete/{id}:
    delete:
      consumes:
      - application/json
      description: ProductDelete elimina un producto
      parameters:
      - description: ID del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductDelete
      tags:
      - Product
  /v1/product/get/{id}:
    get:
      consumes:
      - application/json
      description: ProductGet obtener un producto por ID
      parameters:
      - description: Id del producto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductGet
      tags:
      - Product
  /v1/product/get_all:
    get:
      consumes:
      - application/json
      description: ProductGetAll obtener todos los productos
      parameters:
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      - default: 10
        description: Número de elementos por página
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.ProductResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductGetAll
      tags:
      - Product
  /v1/product/get_by_category/{category_id}:
    get:
      consumes:
      - application/json
      description: ProductGetByCategory obtener un producto por Id de categoria
      parameters:
      - description: ID de la categoria
        in: path
        name: category_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductGetByCategory
      tags:
      - Product
  /v1/product/get_by_code:
    get:
      consumes:
      - application/json
      description: ProductGetByCode obtener un producto por Codigo
      parameters:
      - description: codigo del producto
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductGetByCode
      tags:
      - Product
  /v1/product/get_by_name:
    get:
      consumes:
      - application/json
      description: ProductGetByName obtener un producto por nombre
      parameters:
      - description: nombre del producto
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.ProductResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductGetByName
      tags:
      - Product
  /v1/product/update:
    put:
      consumes:
      - application/json
      description: ProductUpdate edita un producto ya creado
      parameters:
      - description: Información del producto a editar
        in: body
        name: productUpdate
        required: true
        schema:
          $ref: '#/definitions/schemas.ProductUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: ProductUpdate
      tags:
      - Product
  /v1/role/get_all:
    get:
      consumes:
      - application/json
      description: RoleGetAll obtener todos los roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.RoleResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: RoleGetAll
      tags:
      - Role
...
securityDefinitions:
  CookieAuth:
    description: 'Enter your JWT token here. Example: "eyJhbGciOiJIUz..."'
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"